Lots of comments explaining stuff, ClientList.cpp and AgentList.cpp no longer print to the console, AgentMenu.cpp and ClientMenu.cpp now print the lists to the console. I think in ClientList.cpp i changed the delete function a little, not tested

+++ util.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-int countDigits(int n){
+int countDigits(long long n){
     int count = 0;
     while (n != 0)
     {
